requirements.txt
absl-py==2.3.1
astunparse==1.6.3
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudpickle==3.1.1
cryptography==45.0.6
et_xmlfile==2.0.0
fbgemm_gpu==2025.8.18+cu128
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
h5py==3.14.0
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
invoke==2.2.0
iopath==0.1.10
Jinja2==3.1.4
joblib==1.5.1
keras==3.11.2
keras-nightly==3.12.0.dev2025081903
libclang==18.1.1
lightning-utilities==0.15.2
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.3
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pandas-stubs==2.3.0.250703
paramiko==4.0.0
pillow==11.0.0
portalocker==3.2.0
protobuf==6.32.0
psycopg==3.2.9
pyarrow==21.0.0
pycparser==2.22
Pygments==2.19.2
PyNaCl==1.5.0
pyre-extensions==0.0.32
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-triton==3.4.0+gitf7888497
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.43
sshtunnel==0.4.0
sympy==1.13.3
tb-nightly==2.20.0a20250717
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.9.1
tensorflow==2.20.0
termcolor==3.1.0
tf_nightly==2.21.0.dev20250818
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.9.0.dev20250818+cu128
torchmetrics==1.0.3
-e git+https://github.com/pytorch/torchrec@7e9729bab451490e8080eb80760757c8ddd3a0ef#egg=torchrec
torchvision==0.24.0.dev20250818+cu128
tqdm==4.67.1
transformers==4.55.0
triton==3.4.0
types-pytz==2025.2.0.20250809
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
zipp==3.23.0
